requirements.txt
asgiref==3.3.1
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.8
Django==3.1.4
django-ninja==0.10.2
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
packaging==20.8
pathspec==0.8.1
Pillow==8.0.1
pydantic==1.7.3
pyparsing==2.4.7
pytz==2020.5
regex==2020.11.13
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
